fix: Better group API
- `cmd` was split between `cmd` and `errors`.
- Kind of uneven what was brought into the top-level.  This is less
  necessary with good with good docs.

Fixes #40, I hope

BREAKING CHANGE: API organization changed.
